Return pooled model and model list for imputations
After multiple imputation, the single `model` we should return in `model`
attribute  (which is then used by tools such as `modelsummary`) is the
pooled model. This is because that pooled model is the only single
model which recognizes the fact that we are working across multiple
underlying models and thus includes appropriate summary statistics which
are available eg using `broom` methods. We also do not need to return
the full list of models, as 1) the user will already have passed those
models to `mice` or to `marginaleffects` and thus will already have
access to the models, and 2) this will lead to potentially massive
memory usage.